[sociald] Add Facebook photo album synchronisation
This commit extends the Facebook sync adaptor to include photo sync
to a local database.  Currently, it loads photo information for all
photos in all albums, and stores to a local database at:
~/.config/sociald/Images/facebook.db

It does not download any thumbnails or other image data, so the
overhead is low.  The Jolla Gallery Facebook plugin downloads the
image data on demand and updates the image database so that it uses
the locally cached copies when possible.

Contributes to JB#5544