Calendar sync for Facebook, initial work
This commit includes some initial work for importing
Facebook events to the calendar. Right now, it uses
a database to track the list of events from Facebook
and associate them to local mkcal events.

Events are currently imported into the default notebook.
It should be imported into a specific notebook in future
implementations. Even if events cleaning is implemented,
calendar do not support running this cleanup, so it is
untested.