[sociald] Fixes to Twitter notifications. Fixes JB#7596
Instead of posting one notification per mention, we now
group these tweets into one notification. And we don't even
need to register that each tweet was notified in sociald
databases.

Instead, we count the number of new mentions since the last
sync, and we post a notification based on that number.
If there were only one notification, then we display
the content. If there are many of them, we display the
number of mentions. If a notification were already present,
we merge with the existing one instead of posting a new
notification.

This commit also includes a very small fix for Facebook,
that was showing posts not containing media.