[sociald] Update code to use NoUserInteractionPolicy enum.  Contributes to JB#7302
Previously, the QML code in the event view hardcoded the value of the
NoUserInteractionPolicy enum.  This commit changes it to use the enum
from the SignInParameters type.  It also modifies the C++ code to use
that definition of the enum, rather than the one from the SignOn
library, for consistency.

Contributes to JB#7302