[sociald] Initialise clientId parameter statically.  Contributes to JB#7947
The clientId parameter was previously read from the SailfishKeyProvider
every time a new Facebook feed item was posted to the event feed.
This commit statically initialises the clientId to avoid reading the
value from the ini file each time, and to avoid spamming the console
with warnings from the SailfishKeyProvider about end of file.

Contributes to JB#7947