[sociald] Initial work of multiaccount management. Contributes to JB#8122, JB#7966, JB#7965
This commit includes the first part of a rework for the database used
by sociald. It splits the account identifier from syncedData and
include a 1 - * relation, so that many accounts can be related to
one account.

It also introduces UI changes to be able to switch accounts and
interact with the post using one of these accounts.

Cleanups, and some other methods are not yet adapted to the change
as they might need some optimization.