[sociald] Improve the post caching heuristic. Contributes to JB#11132
Previously, the logic to determine whether a post should be cached
included a check that the post included some media (photo, link, or
video).  This commit ensures that if no media is present in the
attachment, the post is still allowed to be cached, which allows
text-only status updates to be cached.

This commit also relaxes the restrictions so that posts which
reference another post (via parent_post_id) or which are posted
to a different wall (target_id) which the user has liked, will
also be allowed and cached.

Contributes to JB#11132