[sociald] Cleanup the child sync adapter on uninit. Contributes to JB#11521
The sociald plugin doesn't get destroyed by Buteo correctly in many
situations, however the uninit() hook does always get called.
As such, we should do as much cleanup as possible in the uninit()
hook to avoid leaking memory and file descriptors.

This commit also ensures that we emit the appropriate setStatus()
signal to tell buteo when we've finished the sync, in error cases.

Contributes to JB#11521