[sociald] Ensure that Google contacts are purged. Contributes to JB#10453
If msyncd crashes before the Google sync process is complete, the
sync timestamp will not be stored for that sync.  If it is the first
sync of a Google account, then no record will exist in the sync
timestamps database that data from that account has been synced.
Thus, when the account is later removed, the sync adapter will not
know that the account existed in the first place, and so won't purge
the previously partially-synced data.

This commit adds a finalCleanup() function which allows sync adapters
to define synchronous purge methods which don't rely on sociald's
sync timestamps to determine which data needs to be purged.

Contributes to JB#10453