[sociald] Replace Facebook feeds view with web browser. Contributes to JB#13153
Previously, we displayed nice Facebook feeds.  This commit hacks
that out and replaces the feeds view icon with a button which opens
up the touch.facebook.com website instead.

It also removes the sync profile xml file for Facebook Posts to ensure
that we don't attempt to sync Post data to device.

Contributes to JB#13153