[sociald] Simplify Facebook contacts sync adapter
Previously, the Facebook contacts sync adapter used a separate db
to store metadata about Facebook contacts, but this metadata is not
necessary, and just adds complexity to the sync adapter.

This commit removes the use of the separate db (provided by
libsocialcache) and instead stores the account metadata into the
QContactOriginMetadata detail of the contacts.

It also simplifies and improves the avatar image handling.