[sociald] Perform delta detection to reduce unnecessary saves. Contributes to JB#13564
This commit adds contact delta detection to determine whether a contact
from a remote source needs to be saved in the database or not.
Previously, we always clobbered the local data with the remote data,
but in many cases this is not necessary because the local data is still
identical to the remote data.

Contributes to JB#13564