[sociald] Fix multiple account contact syncing. Contributes to JB#13564
The previous change to reduce database activity introduced a bug with
multiple account contact syncing.  This commit fixes the error.
This commit also ensures that Facebook friends' birthdays are parsed
correctly (or discarded if they're of an invalid format), to avoid
spurious delta detection causing unnecessary database activity.
Similarly, it ensures that urls (links+websites) are not duplicated,
to avoid spurious deltas.

Contributes to JB#13564