[sociald] Improve avatar sync handling. Contributes to JB#11970
This commit ensures that the improved avatar local-path generation
logic which was recently added to libsocialcache, is used, which
increases the likelihood that remote avatar changes are noticed
and synced.

It also adds logic to the finalCleanup() handler of the Google
contacts sync adapter to remove avatar details from contacts
for whom the avatars could not be downloaded.  This prevents
spurious "empty" avatars from being reported.

Contributes to JB#11970