[sociald] Bail out on Google contacts request error. Contributes to JB#16288
Previously, if a network or other (eg, API request limit) error
occurred during Google contacts sync, it could happen that the JSON
object returned for the error was interpreted by the XML parser as
an (invalid) contact.  In that case, the local-to-remote comparison
function would assume that all of the contacts were removed remotely
and thus need to be removed locally.

This commit ensures that if an error occurs, the sync process bails
out appropriately before attempting to make any changes to the local
database.

Contributes to JB#16288