[sociald] Automatically time out stuck network requests. Contributes to JB#16704
This commit adds code to time out network requests after a certain
period of time.  This ensures that msyncd won't keep the device from
entering suspend if a network request hangs.

It also fixes a problem in the Twitter Home Timeline (Posts) sync
adapter where a network error would cause the finished() handler to
be skipped, which would cause the semaphore to remain undecremented,
which caused the plugin to hang.

Contributes to JB#16704