[sociald] Improve Google avatar handling. Contributes to JB#11970
This commit improves the avatar handling so that avatar images may
be downloaded over multiple sync runs.  This is important if the
user has lots of Google contacts, as they may hit the API call quota
during avatar synchronization during a single sync run.

It also adds account provenance checks to ensure that the sync
adapter works in the case where multiple Google accounts exist on
the device.

Contributes to JB#11970