[sociald] Support the KEY_SYNC_SINCE_DAYS_PAST Buteo option. Contributes to JB#18446
The user is able to select via the account settings UI whether they
wish to see old content in their feeds.  This setting is written to
the KEY_SYNC_SINCE_DAYS_PAST value of the Buteo profile.  This commit
ensures that the tweet / mention / notification sync adaptors filter
out any content which is older than the span specified.

Contributes to JB#18446