[sociald] Use storeSyncContacts to apply avatar changes. Contributes to JB#19718
Google contact avatars are stored into the contacts after the initial
contact data is synchronized.  Previously, this was done via a normal
QContactManager::save() call - but that would result in a change signal
being emitted for the contacts, which might then result in sync being
retriggered (if AlwaysUpToDate / SyncAutomatically is set).

This commit changes the code to use storeSyncContacts which suppresses
the change signal for the sync target, instead.

This commit also doesn't update avatars for contacts which will be
purged due to belonging to a stale account, and it avoids purging
contacts for disabled (but not removed) accounts.

Contributes to JB#19718