[sociald] Ensure that we remove OOB artifacts on purge
The state data will only be initialized during a sync run, not when
an account is removed and the cleanUp() codepath is triggered.
Thus, no cached oob scope will exist in the latter case; we need
to build it manually instead.

This commit does so, to ensure that OOB data for an account is deleted
when the account is removed.