[sociald] Use CleanUpPurge mode for cleanUp() codepath. Contributes to JB#20682
This commit ensures that the plugin is aware that it is performing a
purge for the cleanUp() account-removal case.  In that case, the sync
function will NOT have been called, and so any setup / initialization
which would normally be performed in that function, needs to be done
prior to the purge, and any state finalization needs to be committed
afterward.

In practice, this means that calendar data gets correctly purged from
the databases when the account is deleted.

Contributes to JB#20682