[sociald] Raise the CredentialsNeedUpdate flag for Twitter. Contributes to JB#20716
Previously, we only checked if the error code reported by Twitter was
the "unauthorized" code.  The credentials may also need updating if
the user has revoked access to the account by the application (for
example, by manually updating the password via the web site).

This commit ensures that if the "invalid / expired token" code is
received, that we raise the CredentialsNeedUpdate flag.

Contributes to JB#20716