[buteo-sync-plugins-social] Fix Google Calendar recurrence sync
When a single occurrence of a recurring event is deleted on device,
it is marked as an exception datetime for the original event.

Previously, we only upsynced EXDATE values, not EXDATETIME values,
and similarly we only upsynced RDATE and not RDATETIME values.
This commit ensures that all four are upsynced correctly.

This commit also ensures that when recurrence rules are parsed for
an event, the recurrence information for the event is cleared before
setting the parsed data, to avoid RRULE duplication after sync.