[buteo-sync-plugins-social] Improve logging from signon adapters
This commit improves the logging from the signon adapters, so that
we can more easily see which process is responsible for raising
the CredentialsNeedUpdate flag in user logs without requiring
msyncd to be run with MSYNCD_LOGGING_LEVEL=8 set.

It also updates the error code comparisons from the old,
deprecated SignOn::AuthSession::Error values to the correct
SignOn::Error::ErrorType values.