[buteo-sync-plugins-social] Ignore insubstantial changes in Google contacts. Contributes to MER#1059
Previously, we considered changes to context values (home/work etc)
substantial when syncing contacts to / from Google.  It appears that
the API has recently changed, and context values are now not reported
in the same way from the API.

This commit ensures that we consider those fields to be insubstantial
so as to not contaminate the delta detection algorithm.

Contributes to MER#1059