Avoid overlapping sync windows with Google Calendar.
When we store the sync date associated with each notebook, that sync
date is a millisecond-resolution timestamp.  Unfortunately, mKCal
can only store Notebook sync dates with second-resolution.  This
results in us requesting events with a slightly larger window than
necessary, resulting in Google always returning one event as
modified which we've already synced.

This commit ensures that we don't request that larger window.
This does mean that there is a (sub-second) window in which updates
could be lost if made server-side, which could be solved in the future
by using the SyncToken provided by Google for this purpose.