[buteo-sync-plugins-social] Fix Twitter notification translation strings. Contributes to MER#1064
Some languages have different forms for ownership in sentences like:
"You have 5 new followers"
versus
"You received 5 new mentions"

As such, we need to use separate summary strings for "You" in the
cases where the verb can differ.

This commit also changes the leading character of the body EE string
to be lowercase, as the sentences are "split" between summary and
body (so only the Summary text should be first-letter-capitalised).

Contributes to MER#1064