[buteo-sync-plugins-social] Improve clean sync of Google Calendars. Contributes to MER#1162
This commit ensures that we don't delete the gcal.ini file on package
upgrade, which we previously did (which would then trigger a clean
sync cycle on next synchronization).

It also ensures that when a clean sync does occur (eg, due to error)
that we re-use the same notebook UID after recreating the notebook,
to ensure uid continuity for other applications (eg Calendar App).

Contributes to MER#1162