[buteo-sync-plugins-social] Commit changes after all remote communication succeeds. Contributes to MER#1248
This commit delays any mkcal database activity until after all
communication with the Google Calendar has completed successfully.
This ensures that the database remains in a usable and recoverable
state, and minimises the chance that we need to perform a clean sync
cycle (which can result in lost-local-modifications).
It includes partial-upsync detection via a private extended property.

Contributes to MER#1248